test: reduce CI pipeline duration in debug (#315)
* test: use KMS test server with cert auth for all backend

* test: backend in parallel

* test: reduce test in debug: no package built, no cross-features crates tested

+++ .github/scripts/cargo_build.sh
@@ -10,24 +10,25 @@ set -ex
 
 ROOT_FOLDER=$(pwd)
 
-# First build the Debian and RPM packages. It must come at first since
-# after this step `ckms` and `cosmian_kms_server` are built with custom features flags (fips for example).
-rm -rf target/"$TARGET"/debian
-rm -rf target/"$TARGET"/generate-rpm
-
-if [ -f /etc/redhat-release ]; then
-  cd crate/cli && cargo build --target "$TARGET" --release && cd -
-  cd crate/server && cargo build --target "$TARGET" --release && cd -
-  cargo install --version 0.14.1 cargo-generate-rpm --force
-  cd "$ROOT_FOLDER"
-  cargo generate-rpm --target "$TARGET" -p crate/cli
-  cargo generate-rpm --target "$TARGET" -p crate/server --metadata-overwrite=pkg/rpm/scriptlets.toml
-elif [ -f /etc/lsb-release ]; then
-  cargo install --version 2.4.0 cargo-deb --force
-  cargo deb --target "$TARGET" -p cosmian_kms_cli --variant fips
-  cargo deb --target "$TARGET" -p cosmian_kms_cli
-  cargo deb --target "$TARGET" -p cosmian_kms_server --variant fips
-  cargo deb --target "$TARGET" -p cosmian_kms_server
+if [ "$DEBUG_OR_RELEASE" = "release" ]; then
+  # First build the Debian and RPM packages. It must come at first since
+  # after this step `ckms` and `cosmian_kms_server` are built with custom features flags (fips for example).
+  rm -rf target/"$TARGET"/debian
+  rm -rf target/"$TARGET"/generate-rpm
+  if [ -f /etc/redhat-release ]; then
+    cd crate/cli && cargo build --target "$TARGET" --release && cd -
+    cd crate/server && cargo build --target "$TARGET" --release && cd -
+    cargo install --version 0.14.1 cargo-generate-rpm --force
+    cd "$ROOT_FOLDER"
+    cargo generate-rpm --target "$TARGET" -p crate/cli
+    cargo generate-rpm --target "$TARGET" -p crate/server --metadata-overwrite=pkg/rpm/scriptlets.toml
+  elif [ -f /etc/lsb-release ]; then
+    cargo install --version 2.4.0 cargo-deb --force
+    cargo deb --target "$TARGET" -p cosmian_kms_cli --variant fips
+    cargo deb --target "$TARGET" -p cosmian_kms_cli
+    cargo deb --target "$TARGET" -p cosmian_kms_server --variant fips
+    cargo deb --target "$TARGET" -p cosmian_kms_server
+  fi
 fi
 
 if [ -z "$TARGET" ]; then
@@ -48,6 +49,11 @@ if [ -z "$FEATURES" ]; then
   unset FEATURES
 fi
 
+if [ -z "$KMS_TEST_DB" ]; then
+  echo "Info: KMS_TEST_DB is not set. Forcing sqlite"
+  KMS_TEST_DB="sqlite"
+fi
+
 if [ -z "$SKIP_SERVICES_TESTS" ]; then
   echo "Info: SKIP_SERVICES_TESTS is not set."
   unset SKIP_SERVICES_TESTS
@@ -55,15 +61,18 @@ fi
 
 rustup target add "$TARGET"
 
-# Before building the crates, test crates individually on specific features
-cargo install --version 0.6.31 cargo-hack --force
-crates=("crate/kmip" "crate/client" "crate/server")
-for crate in "${crates[@]}"; do
-  echo "cargo hack on $crate"
-  cd "$crate"
-  cargo hack check --feature-powerset --no-dev-deps
-  cd "$ROOT_FOLDER"
-done
+# Additional tests
+if [ "$DEBUG_OR_RELEASE" = "release" ]; then
+  # Before building the crates, test crates individually on specific features
+  cargo install --version 0.6.31 cargo-hack --force
+  crates=("crate/kmip" "crate/client" "crate/cli" "crate/server")
+  for crate in "${crates[@]}"; do
+    echo "cargo hack on $crate"
+    cd "$crate"
+    cargo hack check --feature-powerset --no-dev-deps
+    cd "$ROOT_FOLDER"
+  done
+fi
 
 echo "Building crate/pkcs11/provider"
 cd crate/pkcs11/provider
@@ -110,26 +119,12 @@ rm -f /tmp/*.json
 
 export RUST_LOG="cosmian_kms_cli=debug,cosmian_kms_server=debug"
 
-# Testing on different databases
-# databases=("redis-findex" "mysql" "sqlite" "postgresql" "sqlite-enc")
-databases=("redis-findex" "sqlite" "postgresql" "sqlite-enc") # for now, do not handle mysql deadlock occurring in CI
-for database in "${databases[@]}"; do
-  # Discard test if skipping options have been passed in `SKIP_SERVICES_TESTS`
-  database_1=$(echo "$database" | cut -d '-' -f 1)
-  database_2=$(echo "$database" | cut -d '-' -f 2)
-  if [ -n "$database_1" ] && [[ "$SKIP_SERVICES_TESTS" == *"$database_1"* ]]; then
-    echo "Skipping tests for $database"
-    continue
-  fi
-  if [ -n "$database_2" ] && [[ "$SKIP_SERVICES_TESTS" == *"$database_2"* ]]; then
-    echo "Skipping tests for $database"
-    continue
-  fi
-  # shellcheck disable=SC2086
-  cargo build --target $TARGET $RELEASE $FEATURES
-  # shellcheck disable=SC2086
-  KMS_TEST_DB="$database" cargo test --target $TARGET $RELEASE $FEATURES --workspace -- --nocapture $SKIP_SERVICES_TESTS
-done
+# shellcheck disable=SC2086
+cargo build --target $TARGET $RELEASE $FEATURES
+
+echo "Database KMS: $KMS_TEST_DB"
+# shellcheck disable=SC2086
+cargo test --target $TARGET $RELEASE $FEATURES --workspace -- --nocapture $SKIP_SERVICES_TESTS
 
 # Uncomment this code to run tests indefinitely
 # counter=1
